test: 테스트를 위해 사용되던 inner클래스 별도 클래스로 분리
